pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "QueryCase"
version = "0.1.0"
description = "A Python package to fetch and semantically search U.S. court cases"
authors = [{ name = "Your Name" }]
dependencies = [
    "requests",
    "pymupdf",
    "faiss-cpu",
    "tqdm",
    "sentence-transformers",
    "transformers",
    "torch"
]
requires-python = ">=3.8"

[project.scripts]
querycase-update = "querycase.update:run"

[tool.setuptools.packages.find]
where = ["."]
include = ["querycase"]
exclude = ["data", "tests"]
